Move UD tagging and CoNLL conversion to separate files.
